remove Column impl on Vec
remove Column impl on Vec to avoid function shadowing
